refactor: move components in landing-page to `components/` folder Refs #6
