+++ src/components/landing/hero-section/bubble-message.tsx
