+++ src/components/landing/hero-section/flip-image.tsx
