+++ src/components/landing/hero-section/snake-beam.tsx
