+++ src/components/landing/project-section/index.tsx
