+++ src/components/landing/project-section/project-list.tsx
