+++ src/components/landing/project-section/project-text.tsx
