+++ src/components/landing/skill-section/cicd-card.tsx
