+++ src/components/landing/skill-section/dynamic-browser-card.tsx
