+++ src/components/landing/skill-section/mobile-card/fly-squid-game.tsx
