+++ src/components/landing/skill-section/mobile-card/mobile-card.tsx
