+++ src/components/landing/skill-section/mobile-card/mobile-device.tsx
