+++ src/components/landing/skill-section/mobile-card/watermelon-game.tsx
