+++ src/components/landing/skill-section/responsive-design-card.tsx
