+++ src/components/landing/skill-section/skill-text.tsx
