+++ src/components/landing/tech-stack-section/teck-stack-text.tsx
